Устранил замечания наставника
