+++ public/index.html
@@ -1,40 +1,51 @@
 <!DOCTYPE html>
 <html lang="en">
   <head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
 
     <title>Big Trip</title>
 
-    <link rel="stylesheet" href="./css/style.css">
+    <link rel="stylesheet" href="./css/style.css" />
   </head>
 
   <body class="page-body">
     <header class="page-header">
-      <div class="page-body__container  page-header__container">
-        <img class="page-header__logo" src="img/logo.png" width="42" height="42" alt="Trip logo">
+      <div class="page-body__container page-header__container">
+        <img
+          class="page-header__logo"
+          src="img/logo.png"
+          width="42"
+          height="42"
+          alt="Trip logo"
+        />
 
         <div class="trip-main">
-          <div class="trip-main__trip-controls  trip-controls">
+          <div class="trip-main__trip-controls trip-controls">
             <div class="trip-controls__filters">
               <h2 class="visually-hidden">Filter events</h2>
               <!-- Фильтры -->
             </div>
           </div>
 
-          <button class="trip-main__event-add-btn  btn  btn--big  btn--yellow" type="button">New event</button>
+          <button
+            class="trip-main__event-add-btn btn btn--big btn--yellow"
+            type="button"
+          >
+            New event
+          </button>
         </div>
       </div>
     </header>
-    <main class="page-body__page-main  page-main">
+    <main class="page-body__page-main page-main">
       <div class="page-body__container">
-          <section class="trip-events">
-            <h2 class="visually-hidden">Trip events</h2>
+        <section class="trip-events">
+          <h2 class="visually-hidden">Trip events</h2>
 
-            <!-- Сортировка -->
+          <!-- Сортировка -->
 
-            <!-- Контент -->
-          </section>
+          <!-- Контент -->
+        </section>
       </div>
     </main>
   </body>
